requirements.txt
appdirs==1.4.3
appnope==0.1.0
bleach==1.5.0
click==6.7
cycler==0.10.0
Cython==0.25.2
decorator==4.0.11
entrypoints==0.2.2
Flask==0.12
Flask-Session==0.3.1
html5lib==0.9999999
ipykernel==4.5.2
ipython==5.2.2
ipython-genutils==0.1.0
ipywidgets==5.2.2
itsdangerous==0.24
Jinja2==2.9.5
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==4.4.0
jupyter-console==5.1.0
jupyter-core==4.3.0
MarkupSafe==0.23
matplotlib==2.0.0
mistune==0.7.3
nbconvert==5.1.1
nbformat==4.2.0
notebook==4.4.1
numpy==1.12.0
olefile==0.44
packaging==16.8
pandocfilters==1.4.1
pexpect==4.2.1
pickleshare==0.7.4
Pillow==4.0.0
prompt-toolkit==1.0.13
ptyprocess==0.5.1
pycocotools==2.0
Pygments==2.2.0
pyparsing==2.1.10
python-dateutil==2.6.0
pytz==2016.10
PyYAML==3.12
pyzmq==16.0.2
qtconsole==4.2.1
requests==2.13.0
simplegeneric==0.8.1
six==1.10.0
terminado==0.6
testpath==0.3
torch==0.1.10.post2
torchvision==0.1.7
tornado==4.4.2
traitlets==4.3.1
wcwidth==0.1.7
Werkzeug==0.11.15
widgetsnbextension==1.2.6
